refactor: Change name cartReducers to cartSlice
